+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- **Number control animation** - Range sliders now animate chart updates when
+  clicking to jump to a new value, but not while dragging
+
 ## [0.5.0] - 2025-12-23
 
 ### Added